fix: add basic error handling to ors snap request and set correct public api snap request size limit
note the documentation states the request size limit to be 5000, but the ors code seems to test for less than the limit. However a local instance can handle 5000 just fine so i'm not quite sure what's up with that. This just makes sure we're on the safe side.

We're also retrying once for a failed snapping request, in case load balancing causes an issue like it has on 2025-05-08